requirements.txt
Flask==2.2.3
flask-marshmallow==0.14.0
flask-sqlalchemy==3.0.3
python-dotenv==1.0.0
Flask-Script==2.0.5
Flask-Migrate==2.5.2
psycopg2==2.9.6
